@obejashaundev: Subiendo cambios para liss

+++ src/routes/index.js
@@ -23,7 +23,7 @@ let credentialValidation = (req, res, next) => {
     let { email, password } = req.body
     try {
         if (!(email && password)) {
-            throw 'All fields must be completed'
+            throw 'Todos los campos deben estar completos'
         }
         next()
     } catch (err) {
@@ -35,7 +35,7 @@ let dataValidation = (req, res, next) => {
     let { names, firstLastname, secondLastname, phone, email, password, rol } = req.body
     try {
         if (!(names && firstLastname && secondLastname && phone && email && password && rol)) {
-            throw 'All fields must be completed'
+            throw 'Todos los campos deben estar completos'
         }
         next()
     } catch (err) {
@@ -48,7 +48,7 @@ let signUpVerifyUserExistence = async (req, res, next) => {
     try {
         let user = await User.findOne({ email })
         if (user) {
-            throw "The user it's already exists"
+            throw "Ya hay un usuario con esta dirección de correo"
         }
         next()
     } catch (err) {
@@ -61,7 +61,7 @@ let signInVerifyUserExistence = async (req, res, next) => {
     try {
         let user = await User.findOne({ email })
         if (!user) {
-            throw "The entered credentials are invalid"
+            throw "Las credenciales introducidas, no son válidas, inténtelo de nuevo"
         }
         req.passwordHash = user.password
         req._id = user._id
@@ -76,9 +76,9 @@ let checkPassword = (req, res, next) => {
         let { password } = req.body
         let passwordHash = req.passwordHash
         let isValid = bcrypt.compareSync(password, passwordHash)
-        let result = '¡Welcome!'
+        let result = '¡Bienvenido!'
         if (!isValid) {
-            result = 'The entered credentials are invalid'
+            result = 'Las credenciales introducidas, no son válidas, inténtelo de nuevo'
             throw result
         }
         req.result = result
@@ -91,12 +91,12 @@ let checkPassword = (req, res, next) => {
 let checkToken = (req, res, next) => {
     try {
         if (!req.headers.authorization) {
-            throw 'Unauthorize request'
+            throw 'Solicitud no autorizada'
         }
         
         let token = req.headers.authorization.split(' ')[1]
         if (!token) {
-            throw 'Unauthorize request'
+            throw 'Solicitud no autorizada'
         }
 
         try {
@@ -116,7 +116,7 @@ let checkToken = (req, res, next) => {
 
 // #region GET paths
 router.get('/', (req, res, next) => {
-    res.send('Hi, this is the Nissan Sales API')
+    res.send('¡Bienvenido, esta es la API de NISSAN SALES APP!')
 })
 router.get('/tasks', (req, res, next) => {
     res.json(new JSONResponse({ data: [
@@ -164,14 +164,14 @@ router.get('/roles', checkToken, async (req, res, next) => {
         let userId = req.userId
         let user = await User.findById(userId).populate('rol')
         if (!user) {
-            throw 'The requesting user could not be found'
+            throw 'El usuario activo no pudo ser encontrado'
         }
         let rolName = user.rol.name
         if (!(rolName.includes(enumRoles.MASTER) || rolName.includes(enumRoles.ADMINISTRADOR))) {
-            throw 'Unauthorize request'
+            throw 'Solicitud no autorizada'
         }
         let data = await Rol.find({ isActive: true, isRemoved: false })
-        let message = 'List of available roles'
+        let message = 'Lista de roles disponibles'
         let response = new JSONResponse({ data, message })
         res.json(response)
     } catch (err) {
@@ -185,13 +185,21 @@ router.get('/users', checkToken, async (req, res, next) => {
         let userId = req.userId
         let user = await User.findById(userId).populate('rol')
         if (!user) {
-            throw 'The requesting user could not be found'
+            throw 'El usuario activo no pudo ser encontrado'
         }
         let rolName = user.rol.name
         if (!(rolName.includes(enumRoles.MASTER) || rolName.includes(enumRoles.ADMINISTRADOR))) {
-            throw 'Unauthorize request'
+            throw 'Solicitud no autorizada'
         }
         let data = await User.find({ isActive: true, isRemoved: false }).populate('rol')
+        if (!rolName.includes(enumRoles.MASTER)) {
+            data = await User.find({ isActive: true, isRemoved: false }).populate({
+                path: 'rol',
+                match: {
+                    name: { $ne: enumRoles.MASTER }
+                }
+            })
+        }
         let message = 'List of users'
         let response = new JSONResponse({ data, message })
         res.json(response)
@@ -207,6 +215,12 @@ router.get('/users', checkToken, async (req, res, next) => {
 router.post('/users', dataValidation, signUpVerifyUserExistence, async (req, res, next) => {
     try {
         let { names, firstLastname, secondLastname, phone, photoBase64, email, password, rol } = req.body
+        
+        let rolDb = await Rol.findOne({ name: rol })
+        if (!rolDb) {
+            throw 'Es necesario definir un rol para este usuario'
+        }
+        
         let photoPath = ''
         if (photoBase64) {
             photoBase64 = photoBase64.replace(/^data:image\/png;base64,/, '')
@@ -214,10 +228,7 @@ router.post('/users', dataValidation, signUpVerifyUserExistence, async (req, res
             photoPath = `/avatars/${imageName}`
             fs.writeFileSync(`..${photoPath}`, photoBase64, 'base64')
         }
-        let rolDb = await Rol.findOne({ name: rol })
-        if (!rolDb) {
-            throw 'Needs define a rol for this user'
-        }
+        
         password = bcrypt.hashSync(password, parseInt(process.env.SALT_ROUNDS))
         let newUser = new User({ rol: rolDb._id, names, firstLastname, secondLastname, phone, photoPath, email, password })
         await newUser.save()
@@ -263,18 +274,18 @@ router.post('/roles', checkToken, async (req, res, next) => {
         let userId = req.userId
         let user = await User.findById(userId)
         if(!user){
-            throw 'The requesting user could not be found'
+            throw 'El usuario activo no pudo ser encontrado'
         }
         if (!user.rol.name == "MASTER"){
-            throw 'Unauthorize request'
+            throw 'Solicitud no autorizada'
         }
         let { name } = req.body
         if(!name){
-            throw 'The name field is not defined'
+            throw 'El campo nombre no ha sido definido'
         }
         let newRol = new Rol({ name })
         await newRol.save()
-        let message = 'The new role was created successfully'
+        let message = 'El nuevo rol se creó satisfactoriamente'
         res.json(new JSONResponse({ message }))
     } catch (err) {
         req.message = err
@@ -290,9 +301,12 @@ router.post('/roles', checkToken, async (req, res, next) => {
 router.delete('/users/:id', checkToken, async (req, res, next) => {
     try {
         let userId = req.params.id
-        let user = await User.findById(userId)
+        let user = await User.findById(userId).populate('rol')
         if (!user) {
-            throw 'The requesting user could not be found'
+            throw 'El usuario seleccionado no pudo ser encontrado'
+        }
+        if(user.rol.name.includes(enumRoles.MASTER)){
+            throw 'El usuario seleccionado no puede ser eliminado debido a que la aplicación necesita de él'
         }
         let response = new JSONResponse({ data: {}, message: {} });
         await User.deleteOne({ _id: userId })
